Example of using OpenAI model
Uses just a simple model to show it, but for our project we need to use a larger one. Keys are NOT committed, but can be found as described in .env.example.

recreate.sh is for updating the conda environment after changing env.yml as noted at https://docs.conda.io/projects/conda/en/latest/user-guide/tasks/manage-environments.html#using-pip-in-an-environment

Dotenv adds a simple class for reading .env.